requirements/base.txt
python-slugify==8.0.2  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower

# Django
# ------------------------------------------------------------------------------
django==4.2.7  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.60.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2023.10  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==5.4.0  # https://github.com/jazzband/django-redis
django-json-widget
django-debug-toolbar
Collectfast==2.2.0


# Theme
django-admin-interface==0.28.3
django-colorfield==0.10.1

#Others
twilio==8.12.0
sendgrid==6.11.0
django-import-export==3.3.6

# Swagger
drf-yasg==1.21.7
django-rest-swagger==2.2.0
djangorestframework-simplejwt==5.3.1

#cors
django-cors-headers==4.3.1

#Xirux
pyodbc==5.0.1
pypyodbc==1.3.6

#Pandas
pandas==2.2.0
